feat: deactivate API call in CreateMoreQuestionsUseCaseTest
The test method `CreateMoreQuestionsUseCaseTest` in the
`CreateMoreQuestionsUseCaseTests.cs` file has been modified
to comment out the actual execution of the
`createMoreQuestionsUseCase.ExecuteAsync` method. This change
effectively deactivates the test, with comments indicating
that it can be reactivated if needed. The reason for
deactivation is that the test makes a real request to the API.