feat: Ensure DateTime properties are stored/retrieved as UTC. It's necessary for postgreSQL db
The changes introduce conversions for several `DateTime` properties
to ensure they are stored and retrieved as UTC in the `AicheckerContext`
class. Specifically, the properties `RequestCreated`, `RequestStart`,
and `RequestEnd` in one entity, and `CpuUsageTimestamp`,
`MemoryUsageTimestamp`, `GpuUsageTimestamp`, and
`GpuTotalMemoryUsageTimestamp` in another entity, are now configured
to convert to UTC when stored and to specify the kind as UTC when
retrieved. This helps maintain time consistency across different
time zones and avoids potential issues with time zone conversions.