feat: Add message parameter to GetRunTable method
The `GetRunTable` method in `ExportPromptRating.cs` was modified to include an additional parameter `message`, which is used to add a message to the table rows. The `modelTable.AddRow` method call within `GetRunTable` was updated to include the `message` parameter, replacing the empty string previously used.

The `ExportPromptRatingUseCase.cs` file was updated to pass the `message` parameter when calling the `GetRunTable` method. The `IExportPromptRating` interface was updated to reflect the new signature of the `GetRunTable` method, adding the `message` parameter. Minor formatting changes were made to the `IExportPromptRating` interface, but no functional changes were introduced in those lines.