refactor: Remove statusHandler from CreateMoreQuestionsUseCase
The `ExecuteAsync` method in `CreateMoreQuestionsUseCase` has been
modified to remove the optional `statusHandler` parameter. This
parameter is no longer part of the method signature.

Correspondingly, the `ExecuteAsync` method in the
`ICreateMoreQuestionsUseCase` interface has also been updated to
remove the `statusHandler` parameter.

Additionally, commented-out code related to handling the status of
the question processing has been removed from the
`CreateMoreQuestionsUseCase` class.