fix: GetRunTable to include message in modelRatings
Refactored the GetRunTable method in ExportPromptRating.cs to change the structure of the modelRatings parameter, now including an additional message field. Updated the loop inside GetRunTable to destructure the modelRatings tuple to include the message field. Modified ExportPromptRatingUseCase.cs to reflect the changes in the GetRunTable method call, passing the message field as part of the modelRatings parameter. Updated IExportPromptRating.cs interface to reflect the new signature of the GetRunTable method, removing the message parameter and including it within the modelRatings list of tuples.