feat: Update model size label from MB to GB in table
The label for the model size in the table has been changed from "Größe des Modells (MB)" (Size of the model in MB) to "Größe des Modells (GB)" (Size of the model in GB). This updates the unit of measurement for the model size from megabytes (MB) to gigabytes (GB) in the `modelDetailsTable`.