feat: Add async test for removing result set in in-memory DB
Added RemoveResultSetAsyncTest in DefaultMethodesRepositoryTests to
verify the removal of a result set asynchronously from the in-memory
database. The test seeds the database with a Result object containing
a ResultSet and SystemResourceUsage entries, then calls
RemoveResultSetAsync and asserts the operation. Added
Microsoft.EntityFrameworkCore.InMemory package version 9.0.0 to
PostgreSqlEFTests.csproj for in-memory database support.