feat: Add "ask" option for UserMessage and clean up code
Introduce a new feature where the UserMessage option can be set to "ask" to prompt the user for input, similar to ResponseFormat and SystemPrompt. If opts.UserMessage is set to "ask", the MultiLineInput method will be called to get the user input for UserMessage. Updated the help text for UserMessage in the SendToLmsVerb class to reflect this change. Additionally, removed redundant empty lines from Application.cs to clean up the code.