feat: Swap message order in CreateMessageForApi method because of cashing the system prompt at the beginning
The CreateMessageForApi method in the ObjectCreationForApi class has been modified to change the order of the messages being created. The first message now has the role "system" and contains the systemPrompt, while the second message has the role "user" and contains the message. This is the reverse of the previous order.